+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-- `status` command
+- `activity` command
 - Option to recurse directories with `download`
 - `cursor` alias to `move`
 - Compatibility table to README